Move tree-based importance tests out of reusable test cases
