+++ optuna/testing/pytest_importance.py
@@ -366,36 +366,3 @@ class MultiObjectiveImportanceEvaluatorTestCase(_BaseImportanceEvaluatorTestCase
 
         param_importance = get_param_importances(study, evaluator=evaluator())
         assert isinstance(param_importance, dict)
-
-
-class TreeBasedImportanceEvaluatorTestCase(_BaseImportanceEvaluatorTestCase):
-    def test_n_trees_of_tree_based_evaluator(
-        self, evaluator: Callable[..., BaseImportanceEvaluator]
-    ) -> None:
-        study = _get_study(seed=0, n_trials=3, is_multi_obj=False)
-        param_importance = evaluator(n_trees=10, seed=0).evaluate(study)
-        param_importance_different_n_trees = evaluator(n_trees=20, seed=0).evaluate(study)
-
-        assert param_importance != param_importance_different_n_trees
-
-    def test_max_depth_of_tree_based_evaluator(
-        self, evaluator: Callable[..., BaseImportanceEvaluator]
-    ) -> None:
-        study = _get_study(seed=0, n_trials=3, is_multi_obj=False)
-        param_importance = evaluator(max_depth=1, seed=0).evaluate(study)
-        param_importance_different_max_depth = evaluator(max_depth=2, seed=0).evaluate(study)
-
-        assert param_importance != param_importance_different_max_depth
-
-    def test_importance_evaluator_seed(
-        self, evaluator: Callable[..., BaseImportanceEvaluator]
-    ) -> None:
-        study = create_study(sampler=RandomSampler(seed=0))
-        study.optimize(_objective, n_trials=3)
-
-        param_importance = evaluator(seed=2).evaluate(study)
-        param_importance_same_seed = evaluator(seed=2).evaluate(study)
-        assert param_importance == param_importance_same_seed
-
-        param_importance_different_seed = evaluator(seed=3).evaluate(study)
-        assert param_importance != param_importance_different_seed
